requirements.txt
aiosqlite==0.20.0
groq==0.11.0
edge-tts==6.1.10
aiohttp==3.9.5
aiofiles==23.2.1
apscheduler==3.10.4
python-dotenv==1.0.1
loguru==0.7.2
mutagen==1.47.0
Pillow==10.3.0
numpy==1.26.4
